engine: extract esm/{equirectangular_tile,render_triangle}.js
