+++ engine/esm/transpiled.js
@@ -144,6 +144,9 @@ import { FitsImage } from "./layers/fits_image.js";
 import { FitsImageJs } from "./layers/fits_image_js.js";
 import { FitsImageTile } from "./layers/fits_image_tile.js";
 
+import { RenderTriangle } from "./render_triangle.js";
+import { EquirectangularTile } from "./equirectangular_tile.js";
+
 
 // wwtlib.ScaleTypes
 
@@ -12560,271 +12563,6 @@ var RenderContext$ = {
 };
 
 registerType("RenderContext", [RenderContext, RenderContext$, null]);
-
-// wwtlib.RenderTriangle
-
-export function RenderTriangle() {
-  this.a = new PositionTexture();
-  this.b = new PositionTexture();
-  this.c = new PositionTexture();
-  this.normal = new Vector3d();
-  this.opacity = 1;
-  this.expansionInPixels = 0.6;
-  this.tileLevel = 0;
-  this._ta = new Vector3d();
-  this._tb = new Vector3d();
-  this._tc = new Vector3d();
-  this._expandedS0 = new Vector2d();
-  this._expandedS1 = new Vector2d();
-  this._expandedS2 = new Vector2d();
-  this.lighting = 1;
-}
-
-RenderTriangle.width = 1024;
-RenderTriangle.height = 768;
-RenderTriangle._contractionInPixels = -0.5;
-RenderTriangle.trianglesRendered = 0;
-RenderTriangle.trianglesCulled = 0;
-RenderTriangle.renderingOn = true;
-RenderTriangle._factor = 1;
-RenderTriangle.cullInside = true;
-RenderTriangle._hw = 0;
-RenderTriangle._qw = 0;
-RenderTriangle._hh = 0;
-RenderTriangle._qh = 0;
-
-RenderTriangle.create = function (a, b, c, img, level) {
-  var temp = new RenderTriangle();
-  temp.a = a.copy();
-  temp.b = b.copy();
-  temp.c = c.copy();
-  temp._texture = img;
-  temp.tileLevel = level;
-  temp.makeNormal();
-  return temp;
-};
-
-RenderTriangle.createWithMiter = function (a, b, c, img, level, expansion) {
-  var temp = new RenderTriangle();
-  temp.expansionInPixels = expansion;
-  temp.a = a.copy();
-  temp.b = b.copy();
-  temp.c = c.copy();
-  temp._texture = img;
-  temp.tileLevel = level;
-  temp.makeNormal();
-  return temp;
-};
-
-RenderTriangle._getMiterPoint = function (p1, p2, p3, edgeOffset) {
-  var edge1 = Vector2d.subtract(p2, p1);
-  var edge2 = Vector2d.subtract(p3, p1);
-  edge1.normalize();
-  edge2.normalize();
-  var dir = Vector2d.create(edge1.x + edge2.x, edge1.y + edge2.y);
-  dir.normalize();
-  var delta = Vector2d.create(edge1.x - edge2.x, edge1.y - edge2.y);
-  var sineHalfAngle = delta.get_length() / 2;
-  var net = Math.min(2, edgeOffset / sineHalfAngle);
-  dir.extend(net);
-  return Vector2d.create(p1.x - dir.x, p1.y - dir.y);
-};
-
-RenderTriangle._miterPoint = function (p1x, p1y, p2x, p2y, p3x, p3y, ExpansionInPixels) {
-  var e1x = p2x - p1x;
-  var e1y = p2y - p1y;
-  var e2x = p3x - p1x;
-  var e2y = p3y - p1y;
-  var length = Math.sqrt(e1x * e1x + e1y * e1y);
-  if (!!length) {
-    e1x /= length;
-    e1y /= length;
-  }
-  length = Math.sqrt(e2x * e2x + e2y * e2y);
-  if (!!length) {
-    e2x /= length;
-    e2y /= length;
-  }
-  var dx = e1x + e2x;
-  var dy = e1y + e2y;
-  length = Math.sqrt(dx * dx + dy * dy);
-  if (!!length) {
-    dx /= length;
-    dy /= length;
-  }
-  var deltax = e1x - e2x;
-  var deltay = e1y - e2y;
-  length = Math.sqrt(deltax * deltax + deltay * deltay);
-  var sineHalfAngle = length / 2;
-  var net = Math.min(2, ExpansionInPixels / sineHalfAngle);
-  dx *= net;
-  dy *= net;
-  return Vector2d.create(p1x - dx, p1y - dy);
-};
-
-RenderTriangle._miterPointOut = function (pntOut, p1x, p1y, p2x, p2y, p3x, p3y, ExpansionInPixels) {
-  var e1x = p2x - p1x;
-  var e1y = p2y - p1y;
-  var e2x = p3x - p1x;
-  var e2y = p3y - p1y;
-  var length = Math.sqrt(e1x * e1x + e1y * e1y);
-  if (!!length) {
-    e1x /= length;
-    e1y /= length;
-  }
-  length = Math.sqrt(e2x * e2x + e2y * e2y);
-  if (!!length) {
-    e2x /= length;
-    e2y /= length;
-  }
-  var dx = e1x + e2x;
-  var dy = e1y + e2y;
-  length = Math.sqrt(dx * dx + dy * dy);
-  if (!!length) {
-    dx /= length;
-    dy /= length;
-  }
-  var deltax = e1x - e2x;
-  var deltay = e1y - e2y;
-  length = Math.sqrt(deltax * deltax + deltay * deltay);
-  var sineHalfAngle = length / 2;
-  var net = Math.min(2, ExpansionInPixels / sineHalfAngle);
-  dx *= net;
-  dy *= net;
-  pntOut.x = p1x - dx;
-  pntOut.y = p1y - dy;
-};
-
-var RenderTriangle$ = {
-  makeNormal: function () {
-    var a = this.a.position.copy();
-    var b = this.b.position.copy();
-    var c = this.c.position.copy();
-    a.normalize();
-    b.normalize();
-    c.normalize();
-    var x = a.x + b.x + c.x;
-    var y = a.y + b.y + c.y;
-    var z = a.z + b.z + c.z;
-    this.normal = Vector3d.create(x / 3, y / 3, z / 3);
-    this.normal.normalize();
-  },
-
-  _checkBackface: function () {
-    var ab = Vector3d.subtractVectors(this._ta, this._tb);
-    var ac = Vector3d.subtractVectors(this._ta, this._tc);
-    var cp = Vector3d.cross(ab, ac);
-    cp.normalize();
-    return cp.z >= 0;
-  },
-
-  draw: function (ctx, wvp) {
-    if (ctx == null) {
-      return;
-    }
-    wvp._transformTo(this.a.position, this._ta);
-    wvp._transformTo(this.b.position, this._tb);
-    wvp._transformTo(this.c.position, this._tc);
-    if (this._checkBackface() === RenderTriangle.cullInside) {
-      RenderTriangle.trianglesCulled++;
-      return;
-    }
-    this._drawTriangle(ctx, this._texture, this._ta.x, this._ta.y, this._tb.x, this._tb.y, this._tc.x, this._tc.y, this.a.tu, this.a.tv, this.b.tu, this.b.tv, this.c.tu, this.c.tv);
-  },
-
-  _drawTriangle: function (ctx, im, x0, y0, x1, y1, x2, y2, sx0, sy0, sx1, sy1, sx2, sy2) {
-    if (!this.intersects(0, RenderTriangle.width, 0, RenderTriangle.height, x0, y0, x1, y1, x2, y2)) {
-      return false;
-    }
-    RenderTriangle._miterPointOut(this._expandedS0, x0, y0, x1, y1, x2, y2, this.expansionInPixels);
-    RenderTriangle._miterPointOut(this._expandedS1, x1, y1, x0, y0, x2, y2, this.expansionInPixels);
-    RenderTriangle._miterPointOut(this._expandedS2, x2, y2, x1, y1, x0, y0, this.expansionInPixels);
-    x0 = this._expandedS0.x;
-    y0 = this._expandedS0.y;
-    x1 = this._expandedS1.x;
-    y1 = this._expandedS1.y;
-    x2 = this._expandedS2.x;
-    y2 = this._expandedS2.y;
-    ctx.save();
-    if (RenderTriangle.renderingOn) {
-      ctx.beginPath();
-      ctx.moveTo(x0, y0);
-      ctx.lineTo(x1, y1);
-      ctx.lineTo(x2, y2);
-      ctx.closePath();
-      ctx.clip();
-    }
-    var denom = sx0 * (sy2 - sy1) - sx1 * sy2 + sx2 * sy1 + (sx1 - sx2) * sy0;
-    var m11 = -(sy0 * (x2 - x1) - sy1 * x2 + sy2 * x1 + (sy1 - sy2) * x0) / denom;
-    var m12 = (sy1 * y2 + sy0 * (y1 - y2) - sy2 * y1 + (sy2 - sy1) * y0) / denom;
-    var m21 = (sx0 * (x2 - x1) - sx1 * x2 + sx2 * x1 + (sx1 - sx2) * x0) / denom;
-    var m22 = -(sx1 * y2 + sx0 * (y1 - y2) - sx2 * y1 + (sx2 - sx1) * y0) / denom;
-    var dx = (sx0 * (sy2 * x1 - sy1 * x2) + sy0 * (sx1 * x2 - sx2 * x1) + (sx2 * sy1 - sx1 * sy2) * x0) / denom;
-    var dy = (sx0 * (sy2 * y1 - sy1 * y2) + sy0 * (sx1 * y2 - sx2 * y1) + (sx2 * sy1 - sx1 * sy2) * y0) / denom;
-    ctx.transform(m11, m12, m21, m22, dx, dy);
-    if (RenderTriangle.renderingOn) {
-      ctx.globalAlpha = this.opacity;
-      if (this.lighting < 1) {
-        ctx.globalAlpha = 1;
-        ctx.fillStyle = 'Black';
-        ctx.fillRect(0, 0, RenderTriangle.width, RenderTriangle.height);
-        ctx.globalAlpha = this.lighting * this.opacity;
-      }
-      ctx.drawImage(im, 0, 0);
-    }
-    ctx.restore();
-    return true;
-  },
-
-  intersects: function (l, r, t, b, x0, y0, x1, y1, x2, y2) {
-    if (x0 > l && x0 < r && y0 > t && y0 < b) {
-      return true;
-    }
-    if (x1 > l && x1 < r && y1 > t && y1 < b) {
-      return true;
-    }
-    if (x2 > l && x2 < r && y2 > t && y2 < b) {
-      return true;
-    }
-    var h4 = RenderTriangle.height * 4;
-    if (this.tileLevel < 4 && ((Math.abs(x0 - x1) > h4) || (Math.abs(y0 - y1) > h4) || (Math.abs(x2 - x1) > h4) || (Math.abs(y2 - y1) > h4) || (Math.abs(x0 - x2) > h4) || (Math.abs(y0 - y2) > h4))) {
-      return false;
-    }
-    return this.lineRectangleIntersect(l, r, t, b, x0, y0, x1, y1) || this.lineRectangleIntersect(l, r, t, b, x1, y1, x2, y2) || this.lineRectangleIntersect(l, r, t, b, x2, y2, x0, y0);
-  },
-
-  lineRectangleIntersect: function (l, r, t, b, x0, y0, x1, y1) {
-    var top_intersection;
-    var bottom_intersection;
-    var toptrianglepoint;
-    var bottomtrianglepoint;
-    var m;
-    var c;
-    m = (y1 - y0) / (x1 - x0);
-    c = y0 - (m * x0);
-    if (m > 0) {
-      top_intersection = (m * l + c);
-      bottom_intersection = (m * r + c);
-    } else {
-      top_intersection = (m * r + c);
-      bottom_intersection = (m * l + c);
-    }
-    if (y0 < y1) {
-      toptrianglepoint = y0;
-      bottomtrianglepoint = y1;
-    } else {
-      toptrianglepoint = y1;
-      bottomtrianglepoint = y0;
-    }
-    var topoverlap;
-    var botoverlap;
-    topoverlap = (top_intersection > toptrianglepoint) ? top_intersection : toptrianglepoint;
-    botoverlap = (bottom_intersection < bottomtrianglepoint) ? bottom_intersection : bottomtrianglepoint;
-    return (topoverlap < botoverlap) && (!((botoverlap < t) || (topoverlap > b)));
-  }
-};
-
-registerType("RenderTriangle", [RenderTriangle, RenderTriangle$, null]);
 
 // wwtlib.ScriptInterface
 
@@ -26226,281 +25964,6 @@ var HealpixTile$ = {
 };
 
 registerType("HealpixTile", [HealpixTile, HealpixTile$, Tile]);
-
-// wwtlib.EquirectangularTile
-
-export function EquirectangularTile() {
-  this._tileDegrees$1 = 0;
-  this._topDown$1 = true;
-  this._subDivisionLevel$1 = 1;
-  Tile.call(this);
-}
-
-EquirectangularTile.create = function (level, x, y, dataset, parent) {
-  var temp = new EquirectangularTile();
-  temp.parent = parent;
-  temp.level = level;
-  temp.tileX = x;
-  temp.tileY = y;
-  temp.dataset = dataset;
-  temp._topDown$1 = !dataset.get_bottomsUp();
-  temp.computeBoundingSphere();
-  return temp;
-};
-
-var EquirectangularTile$ = {
-  computeBoundingSphere: function () {
-    if (!this._topDown$1) {
-      this.computeBoundingSphereBottomsUp();
-      return;
-    }
-    this._tileDegrees$1 = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-    var latMin = (90 - ((this.tileY) * this._tileDegrees$1));
-    var latMax = (90 - (((this.tileY + 1)) * this._tileDegrees$1));
-    var lngMin = ((this.tileX * this._tileDegrees$1) - 180);
-    var lngMax = ((((this.tileX + 1)) * this._tileDegrees$1) - 180);
-    var latCenter = (latMin + latMax) / 2;
-    var lngCenter = (lngMin + lngMax) / 2;
-    this.sphereCenter = this.geoTo3d(latCenter, lngCenter, false);
-    this.topLeft = this.geoTo3d(latMin, lngMin, false);
-    this.bottomRight = this.geoTo3d(latMax, lngMax, false);
-    this.topRight = this.geoTo3d(latMin, lngMax, false);
-    this.bottomLeft = this.geoTo3d(latMax, lngMin, false);
-    var distVect = this.geoTo3d(latMin, lngMin, false);
-    distVect.subtract(this.sphereCenter);
-    this.sphereRadius = distVect.length();
-    this._tileDegrees$1 = lngMax - lngMin;
-  },
-
-  computeBoundingSphereBottomsUp: function () {
-    var tileDegrees = this.dataset.get_baseTileDegrees() / (Math.pow(2, this.level));
-    var latMin = (-90 + (((this.tileY + 1)) * tileDegrees));
-    var latMax = (-90 + ((this.tileY) * tileDegrees));
-    var lngMin = ((this.tileX * tileDegrees) - 180);
-    var lngMax = ((((this.tileX + 1)) * tileDegrees) - 180);
-    var latCenter = (latMin + latMax) / 2;
-    var lngCenter = (lngMin + lngMax) / 2;
-    this.sphereCenter = this.geoTo3d(latCenter, lngCenter, false);
-    this.topLeft = this.geoTo3d(latMin, lngMin, false);
-    this.bottomRight = this.geoTo3d(latMax, lngMax, false);
-    this.topRight = this.geoTo3d(latMin, lngMax, false);
-    this.bottomLeft = this.geoTo3d(latMax, lngMin, false);
-    var distVect = this.topLeft;
-    distVect.subtract(this.sphereCenter);
-    this.sphereRadius = distVect.length();
-    tileDegrees = lngMax - lngMin;
-  },
-
-  createGeometry: function (renderContext) {
-    Tile.prototype.createGeometry.call(this, renderContext);
-    if (renderContext.gl == null) {
-      if (!this.dataset.get_dataSetType() || this.dataset.get_dataSetType() === 1) {
-        this._subDivisionLevel$1 = Math.max(2, (4 - this.level) * 2);
-      }
-    } else {
-      this._subDivisionLevel$1 = 32;
-    }
-    try {
-      for (var i = 0; i < 4; i++) {
-        this._renderTriangleLists[i] = [];
-      }
-      if (!this._topDown$1) {
-        return this._createGeometryBottomsUp$1(renderContext);
-      }
-      var lat, lng;
-      var index = 0;
-      var tileDegrees = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-      var latMin = (90 - ((this.tileY) * tileDegrees));
-      var latMax = (90 - (((this.tileY + 1)) * tileDegrees));
-      var lngMin = ((this.tileX * tileDegrees) - 180);
-      var lngMax = ((((this.tileX + 1)) * tileDegrees) - 180);
-      var tileDegreesX = lngMax - lngMin;
-      var tileDegreesY = latMax - latMin;
-      this.topLeft = this.geoTo3d(latMin, lngMin, false);
-      this.bottomRight = this.geoTo3d(latMax, lngMax, false);
-      this.topRight = this.geoTo3d(latMin, lngMax, false);
-      this.bottomLeft = this.geoTo3d(latMax, lngMin, false);
-      var verts = new Array((this._subDivisionLevel$1 + 1) * (this._subDivisionLevel$1 + 1));
-      var x, y;
-      var textureStep = 1 / this._subDivisionLevel$1;
-      for (y = 0; y <= this._subDivisionLevel$1; y++) {
-        if (y !== this._subDivisionLevel$1) {
-          lat = latMin + (textureStep * tileDegreesY * y);
-        }
-        else {
-          lat = latMax;
-        }
-        for (x = 0; x <= this._subDivisionLevel$1; x++) {
-          if (x !== this._subDivisionLevel$1) {
-            lng = lngMin + (textureStep * tileDegreesX * x);
-          }
-          else {
-            lng = lngMax;
-          }
-          index = y * (this._subDivisionLevel$1 + 1) + x;
-          verts[index] = PositionTexture.createPos(this.geoTo3d(lat, lng, false), x * textureStep, y * textureStep);
-        }
-      }
-      this.triangleCount = this._subDivisionLevel$1 * this._subDivisionLevel$1 * 2;
-      var quarterDivisions = this._subDivisionLevel$1 / 2;
-      var part = 0;
-      if (renderContext.gl == null) {
-        for (var y2 = 0; y2 < 2; y2++) {
-          for (var x2 = 0; x2 < 2; x2++) {
-            index = 0;
-            for (var y1 = (quarterDivisions * y2); y1 < (quarterDivisions * (y2 + 1)); y1++) {
-              for (var x1 = (quarterDivisions * x2); x1 < (quarterDivisions * (x2 + 1)); x1++) {
-                var p1;
-                var p2;
-                var p3;
-                p1 = verts[(y1 * (this._subDivisionLevel$1 + 1) + x1)];
-                p2 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1)];
-                p3 = verts[(y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-                this._renderTriangleLists[part].push(RenderTriangle.create(p1, p3, p2, this.texture, this.level));
-                p1 = verts[(y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-                p2 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1)];
-                p3 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-                this._renderTriangleLists[part].push(RenderTriangle.create(p1, p3, p2, this.texture, this.level));
-              }
-            }
-            part++;
-          }
-        }
-      }
-      else {
-        this._vertexBuffer = tilePrepDevice.createBuffer();
-        tilePrepDevice.bindBuffer(WEBGL.ARRAY_BUFFER, this._vertexBuffer);
-        var f32array = new Float32Array(verts.length * 5);
-        var buffer = f32array;
-        index = 0;
-        var $enum1 = ss.enumerate(verts);
-        while ($enum1.moveNext()) {
-          var pt = $enum1.current;
-          index = this.addVertex(buffer, index, pt);
-        }
-        tilePrepDevice.bufferData(WEBGL.ARRAY_BUFFER, f32array, WEBGL.STATIC_DRAW);
-        for (var y2 = 0; y2 < 2; y2++) {
-          for (var x2 = 0; x2 < 2; x2++) {
-            var ui16array = new Uint16Array(this.triangleCount * 3);
-            var indexArray = ui16array;
-            index = 0;
-            for (var y1 = (quarterDivisions * y2); y1 < (quarterDivisions * (y2 + 1)); y1++) {
-              for (var x1 = (quarterDivisions * x2); x1 < (quarterDivisions * (x2 + 1)); x1++) {
-                indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + x1);
-                indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1);
-                indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-                indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-                indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1);
-                indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-              }
-            }
-            this._indexBuffers[part] = tilePrepDevice.createBuffer();
-            tilePrepDevice.bindBuffer(WEBGL.ELEMENT_ARRAY_BUFFER, this._indexBuffers[part]);
-            tilePrepDevice.bufferData(WEBGL.ELEMENT_ARRAY_BUFFER, ui16array, WEBGL.STATIC_DRAW);
-            part++;
-          }
-        }
-      }
-    }
-    catch ($e2) {
-    }
-    return true;
-  },
-  _createGeometryBottomsUp$1: function (renderContext) {
-    var lat, lng;
-    var index = 0;
-    var tileDegrees = this.dataset.get_baseTileDegrees() / Math.pow(2, this.level);
-    var latMin = (-90 + (((this.tileY + 1)) * tileDegrees));
-    var latMax = (-90 + ((this.tileY) * tileDegrees));
-    var lngMin = ((this.tileX * tileDegrees) - 180);
-    var lngMax = ((((this.tileX + 1)) * tileDegrees) - 180);
-    var tileDegreesX = lngMax - lngMin;
-    var tileDegreesY = latMax - latMin;
-    var verts = new Array((this._subDivisionLevel$1 + 1) * (this._subDivisionLevel$1 + 1));
-    var x, y;
-    var textureStep = 1 / this._subDivisionLevel$1;
-    for (y = 0; y <= this._subDivisionLevel$1; y++) {
-      if (y !== this._subDivisionLevel$1) {
-        lat = latMin + (textureStep * tileDegreesY * y);
-      }
-      else {
-        lat = latMax;
-      }
-      for (x = 0; x <= this._subDivisionLevel$1; x++) {
-        if (x !== this._subDivisionLevel$1) {
-          lng = lngMin + (textureStep * tileDegreesX * x);
-        }
-        else {
-          lng = lngMax;
-        }
-        index = y * (this._subDivisionLevel$1 + 1) + x;
-        verts[index] = PositionTexture.createPos(this.geoTo3d(lat, lng, false), x * textureStep, y * textureStep);
-      }
-    }
-    this.triangleCount = this._subDivisionLevel$1 * this._subDivisionLevel$1 * 2;
-    var quarterDivisions = this._subDivisionLevel$1 / 2;
-    var part = 0;
-    if (renderContext.gl == null) {
-      for (var y2 = 0; y2 < 2; y2++) {
-        for (var x2 = 0; x2 < 2; x2++) {
-          index = 0;
-          for (var y1 = (quarterDivisions * y2); y1 < (quarterDivisions * (y2 + 1)); y1++) {
-            for (var x1 = (quarterDivisions * x2); x1 < (quarterDivisions * (x2 + 1)); x1++) {
-              var p1;
-              var p2;
-              var p3;
-              p1 = verts[(y1 * (this._subDivisionLevel$1 + 1) + x1)];
-              p2 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1)];
-              p3 = verts[(y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-              this._renderTriangleLists[part].push(RenderTriangle.create(p1, p3, p2, this.texture, this.level));
-              p1 = verts[(y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-              p2 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1)];
-              p3 = verts[((y1 + 1) * (this._subDivisionLevel$1 + 1) + (x1 + 1))];
-              this._renderTriangleLists[part].push(RenderTriangle.create(p1, p3, p2, this.texture, this.level));
-            }
-          }
-          part++;
-        }
-      }
-    } else {
-      this._vertexBuffer = tilePrepDevice.createBuffer();
-      tilePrepDevice.bindBuffer(WEBGL.ARRAY_BUFFER, this._vertexBuffer);
-      var f32array = new Float32Array(verts.length * 5);
-      var buffer = f32array;
-      index = 0;
-      var $enum1 = ss.enumerate(verts);
-      while ($enum1.moveNext()) {
-        var pt = $enum1.current;
-        index = this.addVertex(buffer, index, pt);
-      }
-      tilePrepDevice.bufferData(WEBGL.ARRAY_BUFFER, f32array, WEBGL.STATIC_DRAW);
-      for (var y2 = 0; y2 < 2; y2++) {
-        for (var x2 = 0; x2 < 2; x2++) {
-          var ui16array = new Uint16Array(this.triangleCount * 3);
-          var indexArray = ui16array;
-          index = 0;
-          for (var y1 = (quarterDivisions * y2); y1 < (quarterDivisions * (y2 + 1)); y1++) {
-            for (var x1 = (quarterDivisions * x2); x1 < (quarterDivisions * (x2 + 1)); x1++) {
-              indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + x1);
-              indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1);
-              indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-              indexArray[index++] = (y1 * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-              indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + x1);
-              indexArray[index++] = ((y1 + 1) * (this._subDivisionLevel$1 + 1) + (x1 + 1));
-            }
-          }
-          this._indexBuffers[part] = tilePrepDevice.createBuffer();
-          tilePrepDevice.bindBuffer(WEBGL.ELEMENT_ARRAY_BUFFER, this._indexBuffers[part]);
-          tilePrepDevice.bufferData(WEBGL.ELEMENT_ARRAY_BUFFER, ui16array, WEBGL.STATIC_DRAW);
-          part++;
-        }
-      }
-    }
-    return true;
-  }
-};
-
-registerType("EquirectangularTile", [EquirectangularTile, EquirectangularTile$, Tile]);
 
 // wwtlib.GreatCirlceRouteLayer
